Work

+++ source/cli/reghigh/a-extensions.sh
@@ -22,17 +22,17 @@ reghigh::clone(){
     local layer
 
     # It's a regular image, mount the config and layers
-    config="$(echo "$image" | jq -r .config.digest)"
+    config="$(printf "%s" "$image" | jq -r .config.digest)"
 
     dc::logger::info ">>> Mounting config $config into new image $destinationImage"
-    if ! regander::blob::MOUNT "$destinationImage" "$config" "$originImage"; then
+    if ! regander::blob::MOUNT "$destinationImage" "$config" "$originImage" > /dev/null; then
       dc::logger::error "Failed to mount config!"
       exit 1
     fi
 
-    for layer in $(echo "$image" | jq -r .layers[].digest); do
+    for layer in $(printf "%s" "$image" | jq -r .layers[].digest); do
       dc::logger::info ">>> Mounting layer $layer into new image $destinationImage"
-      if ! regander::blob::MOUNT "$destinationImage" "$layer" "$originImage"; then
+      if ! regander::blob::MOUNT "$destinationImage" "$layer" "$originImage" > /dev/null; then
         dc::logger::error "Failed to mount layer!"
         exit 1
       fi
@@ -64,10 +64,8 @@ reghigh::download(){
   local originImage="$1"
   local originReference="$2"
   local destinationFolder="$3"
-  if [ ! -d "$destinationFolder" ]; then
-    mkdir -p "$destinationFolder"
-  fi
 
+  # Get the root manifest
   if ! regander::manifest::GET "$originImage" "$originReference" > "$destinationFolder/manifest.json"; then
     dc::logger::error "Failed to download manifest!"
     exit 1
@@ -76,12 +74,14 @@ reghigh::download(){
   local image
   image="$(cat "$destinationFolder/manifest.json")"
 
+  # Depending on the content-type...
   case $(printf "%s" "$image" | jq -rc .mediaType) in
+  # It's a girl, download the layers
   $MIME_V2_MANIFEST|$MIME_OCI_MANIFEST)
     local config
     local layer
 
-    config="$(echo "$image" | jq -rc .config.digest)"
+    config="$(printf "%s" "$image" | jq -rc .config.digest)"
 
     dc::logger::info "Downloading config $config"
     if ! regander::blob::GET "$originImage" "$config" > "$destinationFolder/config.json"; then
@@ -91,7 +91,7 @@ reghigh::download(){
 
     local x
     x=0
-    for layer in $(echo "$image" | jq -rc .layers[].digest); do
+    for layer in $(printf "%s" "$image" | jq -rc .layers[].digest); do
       x=$(( x + 1 ))
       dc::logger::info "Downloading layer number $x ($layer)"
       if ! regander::blob::GET "$originImage" "$layer" > "$destinationFolder/$x.tar.gz"; then
@@ -99,14 +99,17 @@ reghigh::download(){
         exit 1
       fi
 
-      dc::logger::info "Unpacking layer $x ($layer)"
-      mkdir -p "$destinationFolder/$x"
-      cd "$destinationFolder/$x" > /dev/null || exit 1
-      tar -xzf ../"$x.tar.gz"
-      rm ../"$x.tar.gz"
-      cd - > /dev/null || exit 1
+      if [ "$DC_ARGE_UNPACK" ]; then
+        dc::logger::info "Unpacking layer $x ($layer)"
+        mkdir -p "$destinationFolder/$x"
+        cd "$destinationFolder/$x" > /dev/null || exit 1
+        tar -xzf ../"$x.tar.gz"
+        rm ../"$x.tar.gz"
+        cd - > /dev/null || exit 1
+      fi
     done
     ;;
+  # It's a boy, download the linked manifests
   $MIME_OCI_LIST|$MIME_V2_LIST)
     # It's a list, call clone on every individual image in there
     local item
@@ -117,6 +120,7 @@ reghigh::download(){
       reghigh::download "$originImage" "$(printf "%s" "$item" | jq -rc .digest)" "$destinationFolder/$dest"
     done
     ;;
+  # It's an alien, drop it
   *)
     dc::logger::error "Unsupported manifest type. Not sure what is this: $image"
     exit 1
@@ -125,14 +129,14 @@ reghigh::download(){
 
 }
 
-reghigh::upload-image(){
-  local originFolder="$1"
-  local destinationImage="$2"
-  local destinationReference="$3" # XXX if destref is non-existent, compute a digest
+reghigh::upload(){
+  local destinationImage="$1"
+  local destinationReference="${2:-latest}"
+  local originFolder="$3"
   local image
   local config
   local layer
-  local layers
+  local layers=()
   local layerData
 
   # Pick-up a possible manifest list
@@ -157,28 +161,26 @@ reghigh::upload-image(){
   done
 
   if [ "$islist" ]; then
-    # No, upload the list itself, then exit
+    # Upload the list itself, then exit
     return
   fi
 
-  for layer in "$originFolder"/*; do
-    if [ -d "$layer" ]; then
-      cd "$layer" > /dev/null || exit 1
-      layer="$(basename "$layer")".tar.gz
-      # Pack it
-      tar -cvzf "../$layer" . 2> /dev/null
-      # Upload it
-      if ! layerData=$(regander::blob::PUT "$destinationImage" "$MIME_V2_LAYER" < "../$layer"); then
-        dc::logger::error "Failed to upload layer $layer"
-        exit 1
-      fi
-      cd - > /dev/null || exit 1
-      [ "$layers" ] && layers="$layers,"
-      layers="$layers$layerData"
+  # Not a list - it's a regular image
+
+  # First, consider local tarballs
+  local x
+  x=1
+  while [ -f "$originFolder/$x.tar.gz" ]; do
+    # Upload it
+    if ! layerData=$(regander::blob::PUT "$destinationImage" "$MIME_OCI_GZLAYER" < "$originFolder/$x.tar.gz"); then
+      dc::logger::error "Failed to upload layer $originFolder/$x.tar.gz"
+      exit 1
     fi
+    layers+=("$layerData")
+    x=$(( x + 1 ))
   done
 
-  # Pack the stuff
+  # Now, treat every folder as a subsequent layer
   for layer in "$originFolder"/*; do
     if [ -d "$layer" ]; then
       cd "$layer" > /dev/null || exit 1
@@ -186,28 +188,37 @@ reghigh::upload-image(){
       # Pack it
       tar -cvzf "../$layer" . 2> /dev/null
       # Upload it
-      if ! layerData=$(regander::blob::PUT "$destinationImage" "$MIME_V2_LAYER" < "../$layer"); then
+      if ! layerData=$(regander::blob::PUT "$destinationImage" "$MIME_OCI_GZLAYER" < "../$layer"); then
+        rm "../$layer"
         dc::logger::error "Failed to upload layer $layer"
         exit 1
       fi
+      rm "../$layer"
       cd - > /dev/null || exit 1
-      [ "$layers" ] && layers="$layers,"
-      layers="$layers$layerData"
+      layers+=("$layerData")
     fi
   done
 
   # Now, push the config
-  if ! config=$(regander::blob::PUT "$destinationImage" "$MIME_V2_CONFIG" < "$originFolder/config.json"); then
+  if ! config=$(regander::blob::PUT "$destinationImage" "$MIME_OCI_CONFIG" < "$originFolder/config.json"); then
     dc::logger::error "Failed to upload config $config"
     exit 1
   fi
 
+  local i
+  local partake
+  for i in "${layers[@]}"; do
+    [ "$partake" ] && partake="$partake,"
+    partake="$partake$i"
+  done
+
+  # Finally push the manifest itself, with the rebuilt layer list
   if ! regander::manifest::PUT "$destinationImage" "$destinationReference" < <(printf '{
      "schemaVersion": 2,
      "mediaType": "%s",
      "config": %s,
      "layers": %s
-  }' "$MIME_V2_MANIFEST" "$config" "[$layers]"); then
+  }' "$MIME_OCI_MANIFEST" "$config" "[$partake]"); then
     dc::logger::error "Failed to upload config final image"
     exit 1
   fi
